refactor(ui): extract feed/summary-extract.ts — summary object parsing + classifiers
